Update
Whenever the on/off time is set to zero, the water pump stays on (off-time = 0) or stays off (on-time or on/off-time= 0).

Service mode can be enabled by pressing the Mode button or selecting it in the GUI from the Tools menu. When in service mode the water pump is turned off regardless of its current state and the red LED is blinking. The blocking behavior of the temperature measurment may mess with the blinking delay but nothing to worry about for now.